Fix [circular dependencies] * - thanks to madge
